requirements.txt
emoji
numpy
pandas
Pillow
print_color
psutil
termcolor